Final Fix for mismatched tensor.
The division of layers is not a power of two which was causing the final layer to divide to an irregular number which then could not be unsampled to match. When 960, the height, was divided by 2 for the seventh time the solution comes out 7.5 and rounded to a whole number. The mismatched tensor or up sample would be. . .
torch.Size([32, 112, 14, 10])
torch.Size([32, 112, 15, 10])